Weather condition to play outdoor
Suppose we have a dataset of weather conditions and the corresponding target variable "Play". So using this dataset we need to decide whether we should play or not on a particular day according to the weather conditions. So to solve this problem, we need to follow the below steps:



Convert the given dataset into frequency tables.
Generate a Likelihood table by finding the probabilities of given features.
Now, use the Bayes theorem to calculate the posterior probability.